+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+## [v3.30.0] - 2023-01-29
+
 - Improve `Dnatco` extension
     - Factor out common code in `Dnatco` extension
     - Add `NtC tube` visual. Applicable for structures with NtC annotation
@@ -22,6 +24,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Improve boundary calculation performance
 - Add option to create & include images in state snapshots
 - Fix SSAO artefacts with high bias values
+- Fix SSAO resolution scale parameter handling
 - Improve outlines, visually more stable at different view distances
 
 ## [v3.29.0] - 2023-01-15